Few fixes

+++ db/migrate/20130429123055_transform_played_quiz_students_into_many_to_many.rb
@@ -1,6 +1,6 @@
 class TransformPlayedQuizStudentsIntoManyToMany < ActiveRecord::Migration
   def up
-    remove_column :played_quizzes, :student_ids
+    remove_column :played_quizzes, :student_ids if column_exists?(:played_quizzes, :student_ids)
     create_table :played_quizzes_students, id: false do |t|
       t.integer :student_id
       t.integer :played_quiz_id
